+++ Maze/APPViewController.h
@@ -28,5 +28,6 @@
 @property (strong, nonatomic) NSOperationQueue *queue;
 @property (strong, nonatomic) NSDate *lastUpdateTime;
 @property (assign, nonatomic) BOOL collisionAlertVisible;
+@property (assign, nonatomic) BOOL gameCompleted;
 
 @end